Forecast Diabetes Data-Analysis_Project
We will use Naive Bayes model on the “Prima Indians Diabetes” data set. The model will predict which people are likely to develop diabetes. In particular, all patients here are females at least 21 years old of Pima Indian heritage. The datasets consists of several medical predictor variables.
Dataset:- 
This dataset is originally from the National Institute of Diabetes and Digestive and Kidney Diseases. The objective of the dataset is to diagnostically predict whether or not a patient has diabetes, based on certain diagnostic measurements included in the dataset.

Dataset Description:- 
Several constraints were placed on the selection of these instances from a larger database. The datasets consists of several medical predictor variables and one target variable, Outcome. Predictor variables includes the number of pregnancies the patient has had, their BMI, insulin level, age, and so on.

1.	Pregnancies: Number of times pregnant
2.	Glucose: Plasma glucose concentration a 2 hours in an oral glucose tolerance test
3.	Blood Pressure: Diastolic blood pressure (mm Hg)
4.	Skin Thickness: Triceps skin fold thickness (mm)
5.	Insulin: 2-Hour serum insulin (mu U/ml)
6.	BMI: Body mass index (weight in kg/(height in m)^2)
7.	Diabetes Pedigree Function: Diabetes pedigree function
8.	Age: Age (years)
9.	Outcome: Class variable (0 or 1)